+++ src/main/kotlin/us/appfluent/xwidget/actions/CommandAction.kt
@@ -1,11 +1,10 @@
 package us.appfluent.xwidget.actions
 
+import com.intellij.openapi.actionSystem.ActionUpdateThread
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
-import com.intellij.openapi.project.Project
-import com.intellij.openapi.wm.ToolWindowManager
-import org.jetbrains.plugins.terminal.TerminalToolWindowFactory
-import org.jetbrains.plugins.terminal.TerminalToolWindowManager
+import us.appfluent.xwidget.PluginActionPlaces
+import us.appfluent.xwidget.utils.CommandUtils
 
 
 abstract class CommandAction(private val cmd: String) : AnAction() {
@@ -15,28 +14,12 @@ abstract class CommandAction(private val cmd: String) : AnAction() {
 
     override fun actionPerformed(e: AnActionEvent) {
         val project = e.project ?: return
-        runCommand(project, cmd)
+        val requestFocus = e.place != PluginActionPlaces.BACKGROUND
+        CommandUtils.runCommandInTerminal(project, cmd, requestFocus)
     }
 
-    private fun runCommand(project: Project, cmd: String) {
-        val toolWindow = ToolWindowManager.getInstance(project).getToolWindow(TerminalToolWindowFactory.TOOL_WINDOW_ID)
-        val contentManager = toolWindow?.contentManager
-        val terminal = when (val content = contentManager?.findContent(TAB_NAME)) {
-            null -> {
-                val terminalManager = TerminalToolWindowManager.getInstance(project)
-                terminalManager.createShellWidget(project.basePath, TAB_NAME, true, false)
-            }
-            else -> {
-                val selectRunnable = Runnable {
-                    contentManager.setSelectedContent(content, true)
-                }
-                when (toolWindow.isActive) {
-                    true -> selectRunnable.run()
-                    false -> toolWindow.activate(selectRunnable, true, true)
-                }
-                TerminalToolWindowManager.findWidgetByContent(content)
-            }
-        }
-        terminal?.sendCommandToExecute(cmd)
+    override fun getActionUpdateThread(): ActionUpdateThread {
+        // Choose EDT if interacting with the UI, or BGT for background work
+        return ActionUpdateThread.BGT // or ActionUpdateThread.EDT
     }
 }